pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_chatbot"
version = "0.1.0"
description = "Hybrid AI chatbot with rule-based, RAG, tool-calls, LLM fallback, and human-in-loop"
authors = [{ name="Your Name", email="you@example.com" }]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "fastapi>=0.95.0",
  "uvicorn[standard]>=0.22.0",
  "python-dotenv>=1.0.0",
  "openai>=0.27.0",
  "requests>=2.31.0",
]
